Allocator changes: - Virt manager now work but with some issues - Page fault handler - Phys manager works

+++ src/kernel/include/memory.h
@@ -11,7 +11,7 @@
 #define OFFSET_FROM_BIT(b) 		(b % 0x20)
 
 void* memcpy(void* destination, const void* source, uint16_t num);
-void* memset(void* pointer, int value, uint16_t num);
+void* memset(void* pointer, uint8_t value, uint16_t num);
 int memcmp(const void* firstPointer, const void* secondPointer, uint16_t num);
 void* memmove(void *dest, const void *src, size_t len);
