+++ src/libs/include/memory.h
@@ -8,7 +8,7 @@
 #define SEGOFF2LINE(segoff)     ((SEG(segoff) << 4) + OFF(segoff))
 
 void* memcpy(void* destination, const void* source, uint16_t num);
-void* memset(void* pointer, int value, uint16_t num);
+void* memset(void* pointer, uint8_t value, uint16_t num);
 int memcmp(const void* firstPointer, const void* secondPointer, uint16_t num);
 void* memmove(void *dest, const void *src, size_t len);
